+++ NEWS.md
@@ -1,3 +1,7 @@
+# CHANGES IN DT VERSION 0.12
+
+
+
 # CHANGES IN DT VERSION 0.11
 
 ## NEW FEATURES
@@ -7,6 +11,8 @@
 ## BUG FIXES
 
 - Highlighting when searching with regex (i.e., using `options = list(searchHighlight = TRUE, search = list(regex = TRUE))` for the table) works now (thanks, @lrasmus, #719).
+
+- Updated the AutoFill extension (thanks, @stla @shrektan, #711).
 
 ## MINOR CHANGES